Move dto class move in dto package
